add canvas

+++ src/main/java/edu/ifmo/web/lab3/HitResult.java
@@ -3,23 +3,23 @@ package edu.ifmo.web.lab3;
 import java.io.Serializable;
 
 public class HitResult implements Serializable {
-    private int x;
+    private double x;
     private double y;
     private double r;
     private boolean doesHit;
 
-    public HitResult(int x, double y, double r, boolean doesHit) {
+    public HitResult(double x, double y, double r, boolean doesHit) {
         this.x = x;
         this.y = y;
         this.r = r;
         this.doesHit = doesHit;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
